added notification plus email
