+++ emails/share/html.ejs
@@ -11,6 +11,6 @@
   <body>
     <p>Hi,</p>
     <p><%= name %>, shared <%= count %> documents with you, please click the link below to view and download</p>
-    <a href="http://localhost:3000/shared/<%= id %>">View shared documents</a>
+    <a href="https://d35a2073q06o61.cloudfront.net/shared/<%= id %>_<%= email %>">View shared documents</a>
   </body>
 </html>
