+++ routes/commonUserRoutes.js
@@ -5,5 +5,6 @@ const router = express.Router()
 
 router.route('/articles').get(commonUserController.getAllArticles)
 router.route('/healthTopics').get(commonUserController.getAllHealthTopics)
+router.route('/share').post(commonUserController.getSharedDocs)
 
 module.exports = router
