+++ util/fileUploader.js
@@ -26,12 +26,9 @@ module.exports = async (req, res, next) => {
 
     blobWriter.on('finish', () => {
       // Assembling public URL for accessing the file via HTTP
-      const publicUrl = `https://firebasestorage.googleapis.com/v0/b/${bucket.name}/o/${encodeURI(blob.name)}?alt=media`
-
-      // Return the file name and its public URL
-      // res
-      //   .status(200)
-      //   .send({ fileName: req.file.originalname, fileLocation: publicUrl })
+      const publicUrl = `https://firebasestorage.googleapis.com/v0/b/${
+        bucket.name
+      }/o/${encodeURI(blob.name)}?alt=media`
 
       res.uploadedFile = {
         fileName: req.file.originalname,
